requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-multipart
